requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
mangum==0.17.0
sqlalchemy==2.0.27
python-dotenv==1.0.1
pydantic==2.6.1